Extract LogFactory
